master/agents/parser/parse_backend/requirements_local.txt
# --- AI AGENT & LLM ---
# Cực kì mỏng nhẹ, không cần CUDA hay Torch
google-genai>=0.2.0

# --- STRUCTURED VALIDATION ---
pydantic>=2.9.0

# --- UTILITIES ---
requests
pillow>=10.0.0
python-dotenv>=1.0.0

# --- PDF PROCESSING ---
PyMuPDF>=1.24.0

# --- KNOWLEDGE GRAPH ---
neo4j>=5.20.0
